refactor: allow plugin_out for predeclared_outputs
Refactors the protoc_output rule to allow plugin_out to be set for both
suffix based generation and predeclared outputs. plugin_out defaults
to <rule_dir>/<target_name>. The rule will now fail if the user
specifies a suffix and in combination with the specified output the
file would not be generated under the current Bazel package.