refactor: introduce ProtocPluginInfo provider
The protoc_output rule now relies on plugin information provided with a
ProtocInfoProvider. It contains the executable and a function to derive
the names of the output files that are going to be generated by the
plugin. This adds more flexibility to how the plugin might produce
outputs and avoids repetition that occured e.g. when the suffix was
required to be specified on every protoc_output target that used it.