fix: passing runfiles and options from plugin
This commit changes the ProtocPluginInfo data attribute to a runfiles
attribute that is a depset of files. The protoc_plugin rule collects the
default_runfile from the executables as well as the data files and
provides them via the ProtocPluginInfo runfiles attribute.
Also the default_options of the ProtocPluginInfo are now correctly
passed to the plugin when the action executes. They were missing before.